+++ api/src/main/java/ru/kolobkevic/market/api/dtos/CartItemDto.java
@@ -3,14 +3,16 @@ package ru.kolobkevic.market.api.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class CartItemDto {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private Double price;
-    private Double totalPrice;
+    private BigDecimal price;
+    private BigDecimal totalPrice;
 
     public CartItemDto() {
     }
